Refactor IRGen types
