+++ Sources/IRGen/IULIAInterface.swift
@@ -62,7 +62,7 @@ struct IULIAInterface {
   }
 
   func render(_ functionParameter: Parameter) -> String {
-    return "\(CanonicalType(from: functionParameter.type.rawType)!.rawValue) \(IULIAFunction.mangleIdentifierName(functionParameter.identifier.name))"
+    return "\(CanonicalType(from: functionParameter.type.rawType)!.rawValue) \(Mangler.mangleName(functionParameter.identifier.name))"
   }
 }
