+++ Sources/IRGen/Mangler.swift
@@ -2,11 +2,11 @@
 //  Mangler.swift
 //  IRGen
 //
-//  Created by Schrans, Franklin C P T T on 09/02/2018.
+//  Created by Franklin Schrans on 09/02/2018.
 //
 
 struct Mangler {
-  static func mangledName(_ name: String, enclosingType: String) -> String {
-    return "\(enclosingType)_\(name)"
+  static func mangleName(_ name: String, enclosingType: String? = nil) -> String {
+    return "\(enclosingType ?? "")_\(name)"
   }
 }
